% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate simulated missing mixed-type data.}
\usage{
gendata(
  seed = 1,
  n = 300,
  p = 50,
  type = c("homonorm", "heternorm", "pois", "bino", "norm_pois", "pois_bino", "npb"),
  q = 6,
  mis_vec = 0.3,
  rho = 1,
  n_bin = 1,
  mechanism = "MCAR"
)
}
\arguments{
\item{seed}{A non-negative integer, the random seed.}

\item{n}{A positive integer, the sample size.}

\item{p}{A positive integer, the variable dimension.}

\item{type}{A stering, the variable types to be generated.}

\item{q}{A positive integer, the number of factors.}

\item{mis_vec}{A real or real vector with values between 0 and 1, the missing rate of each variable.}

\item{rho}{A positive real or postive vector with two components, the magnitude of loading matrix.}

\item{n_bin}{A positive integer, the number of trails in generating binomial variables, default as 1.}

\item{mechanism}{A string, the missing mechanism.}
}
\description{
The function \code{gendata} is used to generate simulated missing mixed-type data
}
\examples{
datList <- gendata(seed=1)

}
