% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherMethods.R
\name{LFM.XP}
\alias{LFM.XP}
\title{Missing data imputation using the methods (linear factor models)  in Xiong (2021).}
\usage{
LFM.XP(Xmis, q, prop.weighted = TRUE)
}
\arguments{
\item{Xmis}{A n-by-p matrix with missing values, the missing data.}

\item{q}{A positive integer, the number of factors.}

\item{prop.weighted}{A logical value, if it is TRUE, the LFM-PR is used, otherwise, the LFM-BR is used.}
}
\description{
The function \code{LFM.JMS} imputes missing data using the methods (linear factor models)  in Jin (2021).
}
\examples{
datList <- gendata(seed=1)
hX <- LFM.XP(datList$Xmis, q=4)
}
